created dedicated directory for types
