+++ src/types/permissions.ts
@@ -1,0 +1,1 @@
+export type PermissionCheckTypes = "ALL" | "ANY";
